# server/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Prysm Phase 5 API",
    "version": "0.5.0"
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/suspects/top": {
      "get": {
        "summary": "Rank all observed people at an explicit cutoff using the selected engine",
        "description": "{data:[RankingEntry with subjectId],population,cutoffAt,page}. No backend scoring. Clearance filtering may return fewer than limit. Access: subject:read, minimum rank 2",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "cutoffAt",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{data:[RankingEntry with subjectId],population,cutoffAt,page}. No backend scoring. Clearance filtering may return fewer than limit",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "{data:[RankingEntry with subjectId],population,cutoffAt,page}. No backend scoring. Clearance filtering may return fewer than limit"
                }
              }
            }
          },
          "400": {
            "description": "Standard error: {error:{code,message,requestId,details?}}"
          },
          "401": {
            "description": "Standard error: {error:{code,message,requestId,details?}}"
          },
          "403": {
            "description": "Standard error: {error:{code,message,requestId,details?}}"
          },
          "404": {
            "description": "Standard error: {error:{code,message,requestId,details?}}"
          },
          "409": {
            "description": "Standard error: {error:{code,message,requestId,details?}}"
          },
          "413": {
            "description": "Standard error: {error:{code,message,requestId,details?}}"
          },
          "429": {
            "description": "Standard error: {error:{code,message,requestId,details?}}"
          },
          "500": {
            "description": "Standard error: {error:{code,message,requestId,details?}}"
          },
          "502": {
            "description": "Standard error: {error:{code,message,requestId,details?}}"
          },
          "503": {
            "description": "Standard error: {error:{code,message,requestId,details?}}"
          }
        }
      }
    },
    "/dashboard/top-suspects": {
      "get": {
        "summary": "Alias for the same canonical ranking",
        "description": "Same as GET /suspects/top; cutoffAt is required. Access: subject:read, minimum rank 2",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "cutoffAt",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Same as GET /suspects/top; cutoffAt is required",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Same as GET /suspects/top; cutoffAt is required"
                }
              }
            }
          },
          "400": {
            "description": "Standard error: {error:{code,message,requestId,details?}}"
          },
          "401": {
            "description": "Standard error: {error:{code,message,requestId,details?}}"
          },
          "403": {
            "description": "Standard error: {error:{code,message,requestId,details?}}"
          },
          "404": {
            "description": "Standard error: {error:{code,message,requestId,details?}}"
          },
          "409": {
            "description": "Standard error: {error:{code,message,requestId,details?}}"
          },
          "413": {
            "description": "Standard error: {error:{code,message,requestId,details?}}"
          },
          "429": {
            "description": "Standard error: {error:{code,message,requestId,details?}}"
          },
          "500": {
            "description": "Standard error: {error:{code,message,requestId,details?}}"
          },
          "502": {
            "description": "Standard error: {error:{code,message,requestId,details?}}"
          },
          "503": {
            "description": "Standard error: {error:{code,message,requestId,details?}}"
          }
        }
      }
    },
    "/investigations/{id}/intelligence": {
      "get": {
        "summary": "Retrieve the latest successful current-contract result",
        "description": "{runId,result:Intelligence}; 409 ANALYSIS_REQUIRED or ANALYSIS_STALE; legacy v1 results require reanalysis. Access: investigation:read, minimum rank 2",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{runId,result:Intelligence}; 409 ANALYSIS_REQUIRED or ANALYSIS_STALE; legacy v1 results require reanalysis",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "{runId,result:Intelligence}; 409 ANALYSIS_REQUIRED or ANALYSIS_STALE; legacy v1 results require reanalysis"
                }
              }
            }
          },
          "400": {
            "description": "Standard error: {error:{code,message,requestId,details?}}"
          },
          "401": {
            "description": "Standard error: {error:{code,message,requestId,details?}}"
          },
          "403": {
            "description": "Standard error: {error:{code,message,requestId,details?}}"
          },
          "404": {
            "description": "Standard error: {error:{code,message,requestId,details?}}"
          },
          "409": {
            "description": "Standard error: {error:{code,message,requestId,details?}}"
          },
          "413": {
            "description": "Standard error: {error:{code,message,requestId,details?}}"
          },
          "429": {
            "description": "Standard error: {error:{code,message,requestId,details?}}"
          },
          "500": {
            "description": "Standard error: {error:{code,message,requestId,details?}}"
          },
          "502": {
            "description": "Standard error: {error:{code,message,requestId,details?}}"
          },
          "503": {
            "description": "Standard error: {error:{code,message,requestId,details?}}"
          }
        }
      }
    },
    "/investigations/{id}/conversation": {
      "get": {
        "summary": "Read your conversation history for an accessible analyzed case",
        "description": "{data:[{id,conversationId,question,answer,sources,createdAt,contextManifest}],page}; newest first. Access: chat:authorized, minimum rank 2",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{data:[{id,conversationId,question,answer,sources,createdAt,contextManifest}],page}; newest first",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "{data:[{id,conversationId,question,answer,sources,createdAt,contextManifest}],page}; newest first"
                }
              }
            }
          },
          "400": {
            "description": "Standard error: {error:{code,message,requestId,details?}}"
          },
          "401": {
            "description": "Standard error: {error:{code,message,requestId,details?}}"
          },
          "403": {
            "description": "Standard error: {error:{code,message,requestId,details?}}"
          },
          "404": {
            "description": "Standard error: {error:{code,message,requestId,details?}}"
          },
          "409": {
            "description": "Standard error: {error:{code,message,requestId,details?}}"
          },
          "413": {
            "description": "Standard error: {error:{code,message,requestId,details?}}"
          },
          "429": {
            "description": "Standard error: {error:{code,message,requestId,details?}}"
          },
          "500": {
            "description": "Standard error: {error:{code,message,requestId,details?}}"
          },
          "502": {
            "description": "Standard error: {error:{code,message,requestId,details?}}"
          },
          "503": {
            "description": "Standard error: {error:{code,message,requestId,details?}}"
          }
        }
      }
    },
...
  }
}